+++ changelog.md
@@ -7,6 +7,7 @@
 * fixed: WooCommerce 2.6 compatibility (deprecated notice on credit card form)
 * added: basic logging for eWAY transactions (WooCommerce, WP eCommerce, Events Manager)
 * changed: remove dependency on WP eCommerce deprecated function for checkout field list
+* changed: use `wp_remote_retrieve_body()` instead of array access to get eWAY response (WP4.6 compatibility)
 
 ### 3.5.0, 2015-10-10