+++ changelog.md
@@ -5,7 +5,7 @@
 ### 3.5.1, soon...
 
 * fixed: WooCommerce 2.6 compatibility (deprecated notice on credit card form)
-* added: basic logging for eWAY transactions (WooCommerce, WP eCommerce, Events Manager)
+* added: basic logging for eWAY transactions
 * changed: remove dependency on WP eCommerce deprecated function for checkout field list
 * changed: use `wp_remote_retrieve_body()` instead of array access to get eWAY response (WP4.6 compatibility)
 * changed: improved accessibility for credit card forms, admin settings forms